Separate local package for ui components
