+++ packages/ui_components/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 1.0.0
+
+* Initial release: CVCardContainer, LinkIcon and text widgets.
